audio-transcription/requirements.txt
openai-whisper>=20231117
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
huggingface-hub>=0.19.0
